strips out hooks for reuse
